externalized command handling
