+++ remoting/dolphin-remoting-client/src/main/groovy/org/opendolphin/core/client/comm/AbstractClientConnector.groovy
@@ -17,10 +17,6 @@ package org.opendolphin.core.client.comm
 import groovy.transform.CompileStatic
 import groovy.util.logging.Log
 import org.codehaus.groovy.runtime.StackTraceUtils
-import org.opendolphin.core.Attribute
-import org.opendolphin.core.PresentationModel
-import org.opendolphin.core.Tag
-import org.opendolphin.core.client.ClientAttribute
 import org.opendolphin.core.client.ClientDolphin
 import org.opendolphin.core.client.ClientModelStore
 import org.opendolphin.core.client.ClientPresentationModel
@@ -33,7 +29,6 @@ import java.util.logging.Level
 @Log
 abstract class AbstractClientConnector implements ClientConnector {
 
-    boolean strictMode = true // disallow value changes that are based on improper old values
     Codec codec
     UiThreadHandler uiThreadHandler // must be set from the outside - toolkit specific
 
@@ -41,8 +36,23 @@ abstract class AbstractClientConnector implements ClientConnector {
 
     Closure onException;
 
+    ClientResponseHandler responseHandler;
+
     protected final ClientDolphin clientDolphin
     protected final ICommandBatcher commandBatcher
+
+
+    /** The named command that waits for pushes on the server side */
+    NamedCommand  pushListener   = null;
+    /** The signal command that publishes a "release" event on the respective bus */
+    SignalCommand releaseCommand = null;
+
+    /** whether listening for push events should be done at all. */
+    protected boolean pushEnabled = false;
+
+    /** whether we currently wait for push events (internal state) and may need to release */
+    protected boolean waiting = false;
+
 
     AbstractClientConnector(ClientDolphin clientDolphin) {
         this(clientDolphin, null)
@@ -51,6 +61,7 @@ abstract class AbstractClientConnector implements ClientConnector {
     AbstractClientConnector(ClientDolphin clientDolphin, ICommandBatcher commandBatcher) {
         this.clientDolphin = clientDolphin
         this.commandBatcher = commandBatcher ?: new CommandBatcher()
+        this.responseHandler = new ClientResponseHandler(clientDolphin);
 
         // see https://issues.apache.org/jira/browse/GROOVY-7233 and https://issues.apache.org/jira/browse/GROOVY-5438
         def log = log;
@@ -182,169 +193,10 @@ abstract class AbstractClientConnector implements ClientConnector {
         doExceptionSafe doInside
     }
 
-    def handle(Command serverCommand) {
-        log.severe "C: cannot handle unknown command '$serverCommand'"
-    }
 
-    Map handle(DataCommand serverCommand) {
-        return serverCommand.data
-    }
-
-    ClientPresentationModel handle(DeletePresentationModelCommand serverCommand) {
-        ClientPresentationModel model = clientDolphin.findPresentationModelById(serverCommand.pmId)
-        if (!model) return null
-        clientModelStore.delete(model)
-        return model
-    }
-
-    ClientPresentationModel handle(DeleteAllPresentationModelsOfTypeCommand serverCommand) {
-        clientDolphin.deleteAllPresentationModelsOfType(serverCommand.pmType)
-        return null // we cannot really return a single pm here
-    }
-
-    @CompileStatic
-    ClientPresentationModel handle(CreatePresentationModelCommand serverCommand) {
-        if (clientModelStore.containsPresentationModel(serverCommand.pmId)) {
-            throw new IllegalStateException("There already is a presentation model with id '$serverCommand.pmId' known to the client.")
-        }
-        List<ClientAttribute> attributes = []
-        for (attr in serverCommand.attributes) {
-            ClientAttribute attribute = new ClientAttribute(
-                attr.propertyName.toString(),
-                attr.value,
-                attr.qualifier?.toString(),
-                attr.tag ? Tag.tagFor[(String) attr.tag] : Tag.VALUE)
-            if(attr.id?.toString()?.endsWith('S')) {
-                attribute.id = attr.id
-            }
-            attribute.baseValue = attr.baseValue
-            attributes << attribute
-        }
-        ClientPresentationModel model = new ClientPresentationModel(serverCommand.pmId, attributes)
-        model.presentationModelType = serverCommand.pmType
-        if (serverCommand.clientSideOnly) {
-            model.clientSideOnly = true
-        }
-        clientModelStore.add(model)
-        clientDolphin.updateQualifiers(model)
-        return model
-    }
-
-    ClientPresentationModel handle(ValueChangedCommand serverCommand) {
-        Attribute attribute = clientModelStore.findAttributeById(serverCommand.attributeId)
-        if (!attribute) {
-            log.warning "C: attribute with id '$serverCommand.attributeId' not found, cannot update old value '$serverCommand.oldValue' to new value '$serverCommand.newValue'"
-            return null
-        }
-        if (attribute.value?.toString() == serverCommand.newValue?.toString()) {
-            return null
-        }
-        if (strictMode && attribute.value?.toString() != serverCommand.oldValue?.toString()) {
-            // todo dk: think about sending a RejectCommand here to tell the server about a possible lost update
-            log.warning "C: attribute with id '$serverCommand.attributeId' and value '$attribute.value' cannot be set to new value '$serverCommand.newValue' because the change was based on an outdated old value of '$serverCommand.oldValue'."
-            return null
-        }
-        log.info "C: updating '$attribute.propertyName' id '$serverCommand.attributeId' from '$attribute.value' to '$serverCommand.newValue'"
-        attribute.value = serverCommand.newValue
-        return null // this command is not expected to be sent explicitly, so no pm needs to be returned
-    }
-
-    ClientPresentationModel handle(SwitchPresentationModelCommand serverCommand) {
-        def switchPm = clientModelStore.findPresentationModelById(serverCommand.pmId)
-        if (!switchPm) {
-            log.warning "C: switch pm with id '$serverCommand.pmId' not found, cannot switch"
-            return null
-        }
-        def sourcePm = clientModelStore.findPresentationModelById(serverCommand.sourcePmId)
-        if (!sourcePm) {
-            log.warning "C: source pm with id '$serverCommand.sourcePmId' not found, cannot switch"
-            return null
-        }
-        switchPm.syncWith sourcePm                  // ==  clientDolphin.apply sourcePm to switchPm
-        return (ClientPresentationModel) switchPm
-    }
-
-    ClientPresentationModel handle(InitializeAttributeCommand serverCommand) {
-        def attribute = new ClientAttribute(serverCommand.propertyName, serverCommand.newValue, serverCommand.qualifier, serverCommand.tag)
-
-        // todo: add check for no-value; null is a valid value
-        if (serverCommand.qualifier) {
-            def copies = clientModelStore.findAllAttributesByQualifier(serverCommand.qualifier)
-            if (copies) {
-                if (null == serverCommand.newValue) {
-                    attribute.value = copies.first()?.value
-                } else {
-                    copies.each { attr ->
-                        attr.value = attribute.value
-                    }
-                }
-            }
-        }
-        ClientPresentationModel presentationModel = null
-        if (serverCommand.pmId) presentationModel = clientModelStore.findPresentationModelById(serverCommand.pmId)
-        // here we could have a pmType conflict and we may want to throw an Exception...
-        // if there is no pmId, it is most likely an error and CreatePresentationModelCommand should have been used
-        if (!presentationModel) {
-            presentationModel = new ClientPresentationModel(serverCommand.pmId, [])
-            presentationModel.setPresentationModelType(serverCommand.pmType)
-            clientModelStore.add(presentationModel)
-        }
-        // if we already have the attribute, just update the value
-        def existingAtt = presentationModel.getAt(serverCommand.propertyName, serverCommand.tag)
-        if (existingAtt) {
-            existingAtt.value = attribute.value
-        } else {
-            clientDolphin.addAttributeToModel(presentationModel, attribute)
-        }
-        clientDolphin.updateQualifiers(presentationModel)
-        return presentationModel // todo dk: check and test
-    }
-
-    ClientPresentationModel handle(SavedPresentationModelNotification serverCommand) {
-        if (!serverCommand.pmId) return null
-        ClientPresentationModel model = clientModelStore.findPresentationModelById(serverCommand.pmId)
-        if (null == model) {
-            log.warning("model with id '$serverCommand.pmId' not found, cannot rebase")
-            return null
-        }
-        model.attributes*.rebase() // rebase sends update command if needed through PCL
-        return model
-    }
-
-    ClientPresentationModel handle(PresentationModelResetedCommand serverCommand) {
-        if (!serverCommand.pmId) return null
-        PresentationModel model = clientModelStore.findPresentationModelById(serverCommand.pmId)
-        // reset locally first
-        if (!model) return null
-        model.attributes*.reset()
-        return model
-    }
-
-    ClientPresentationModel handle(AttributeMetadataChangedCommand serverCommand) {
-        ClientAttribute attribute = clientModelStore.findAttributeById(serverCommand.attributeId)
-        if (!attribute) return null
-        attribute[serverCommand.metadataName] = serverCommand.value
-        return null
-    }
-
-    ClientPresentationModel handle(CallNamedActionCommand serverCommand) {
-        clientDolphin.send(serverCommand.actionName)
-        return null
-    }
 
 
     //////////////////////////////// push support ////////////////////////////////////////
-
-    /** The named command that waits for pushes on the server side */
-    NamedCommand  pushListener   = null;
-    /** The signal command that publishes a "release" event on the respective bus */
-    SignalCommand releaseCommand = null;
-
-    /** whether listening for push events should be done at all. */
-    protected boolean pushEnabled = false;
-
-    /** whether we currently wait for push events (internal state) and may need to release */
-    protected boolean waiting = false;
 
     /** listens for the pushListener to return. The pushListener must be set and pushEnabled must be true. */
     public void listen() {
@@ -386,5 +238,62 @@ abstract class AbstractClientConnector implements ClientConnector {
     @Override
     boolean isPushEnabled() {
         return this.pushEnabled;
+    }
+
+    def handle(Command serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    Map handle(DataCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(DeletePresentationModelCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(DeleteAllPresentationModelsOfTypeCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    @CompileStatic
+    ClientPresentationModel handle(CreatePresentationModelCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(ValueChangedCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(SwitchPresentationModelCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(InitializeAttributeCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(SavedPresentationModelNotification serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(PresentationModelResetedCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(AttributeMetadataChangedCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    ClientPresentationModel handle(CallNamedActionCommand serverCommand) {
+        return responseHandler.handle(serverCommand);
+    }
+
+    boolean getStrictMode() {
+        return this.responseHandler.strictMode;
+    }
+
+    void setStrictMode(boolean strictMode) {
+        this.responseHandler.strictMode = strictMode;
     }
 }
